Remove BigIntegerField change

+++ jcasts/episodes/models.py
@@ -128,8 +128,8 @@ class Episode(models.Model):
     link: str | None = models.URLField(max_length=2083, null=True, blank=True)
 
     episode_type: str = models.CharField(max_length=30, default="full")
-    episode: int | None = models.BigIntegerField(null=True, blank=True)
-    season: int | None = models.BigIntegerField(null=True, blank=True)
+    episode: int | None = models.IntegerField(null=True, blank=True)
+    season: int | None = models.IntegerField(null=True, blank=True)
 
     cover_url: str | None = models.URLField(max_length=2083, null=True, blank=True)
